backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
requests
pydantic-settings
python-jose[cryptography]
